refactor: use getUserId for getting user ID instead of useUser
