+++ project/app/(dashboard)/projects/[id]/page.tsx
@@ -1,7 +1,7 @@
 "use server";
 
 import { DashboardContent } from "@/components/layouts/dashboard/dashboard-content";
-import { KanbanBoard } from "@/components/sections/project-view/kanban-board";
+import KanbanBoard from "@/components/sections/project-view/kanban-board";
 import ProjectActionButtons from "@/components/sections/project-view/project-action-buttons";
 import { getProject } from "@/lib/actions/projects";
